+++ HDF5Api/H5DataSetWriter.cs
@@ -15,13 +15,13 @@ public static class H5DataSetWriter
         // NOTE: we're only interested in creating a data set currently, not opening an existing one
 
         // Single dimension (rank 1), unlimited length, chunk size.
-        using var memorySpace = H5SAdapter.CreateSimple(new Dimension((ulong)chunkSize));
+        using var memorySpace = H5SAdapter.CreateSimple(new Dimension(chunkSize));
 
         // Create a dataset-creation property list
         using var propertyList = H5PAdapter.Create(H5P.DATASET_CREATE);
 
         // Enable chunking. From the user guide: "HDF5 requires the use of chunking when defining extendable datasets."
-        propertyList.SetChunk(1, (ulong)chunkSize);
+        propertyList.SetChunk(1, chunkSize);
 
         if (compressionLevel > 0)
         {
